requirements.txt
Django==2.2.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
coverage==4.5.4
PyYAML==5.1.2
drf-yasg==1.16.1
PyJWT==1.7.1
packaging==19.2
flake8==3.7.8
django-cors-headers==3.1.0
psycopg2-binary>=2.7.5,<2.8.0
serum==5.1.0
